+++ src/main/kotlin/org/meogo/domain/user/facade/UserFacade.kt
@@ -1,8 +1,8 @@
 package org.meogo.domain.user.facade
 
 import org.meogo.domain.user.domain.User
+import org.meogo.domain.user.domain.UserRepository
 import org.meogo.domain.user.exception.UserNotFoundException
-import org.meogo.domain.user.repository.UserRepository
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Component
 import java.util.UUID
